[fix] Fix useless warnings of 'deleteAll' peakgroups

+++ src/gui/mzroll/tabledockwidget.cpp
@@ -628,6 +628,8 @@ QList<shared_ptr<PeakGroup>> TableDockWidget::getGroups()
 bool TableDockWidget::deleteAll()
 {
   
+  if (!topLevelGroupCount())
+    return false;
   auto continueDeletion = deleteAllgroupsWarning();
   if (!continueDeletion)
     return false;
